fix for issue 198 (setup.py chokes when no arguments given)

+++ setup.py
@@ -228,7 +228,7 @@ if has_cython:
     # recompile netCDF4.pyx
     extensions = [Extension("netCDF4",["netCDF4.pyx"],libraries=libs,library_dirs=lib_dirs,include_dirs=inc_dirs,runtime_library_dirs=lib_dirs)]
     # remove netCDF4.c file if it exists, so cython will recompile netCDF4.pyx.
-    if sys.argv[1] == 'build' and os.path.exists('netCDF4.c'):
+    if len(sys.argv) >= 2 and sys.argv[1] == 'build' and os.path.exists('netCDF4.c'):
         os.remove('netCDF4.c')
     # this determines whether renameGroup method will work.
     has_rename_grp = check_has_rename_grp(inc_dirs)
